Add AbstractChain
